Add assignment operator to TestBar test util class.
I added a an assignment operator to the class, and I added a field
named `magic` that I set to certain value during
construction. Only an object that has been constructed is likely to
have that magic value properly set.

An assignment precondition is that the left-hand side is valid.
The assignment operators I added assert that the `magic` member of
the lhs has the correct value.