Move local static for device count to a host-only function
A future release of NVC++ will report an error for local static variables
that are defined in device functions and host/device functions, even if
the variable is defined within a host-only region of the function.  To
avoid errors, move the local static variable that stores the device count
from cub::DeviceCount() to a separate host-only function.

Fix #1217