Remove CUB_PTX_ARCH, etc from cub::DeviceSpmv.
`cub::DeviceSpmv` has not been ported to use cub::detail::ptx_dispatch
and uses an older dispatch mechanism that is incompatible with nvc++.

It was not ported because there are no tests and thus no way to verify
whether or not the ported version works. Rather than risk breaking
user code, we'll port this later, after we add tests. NV HPC doesn't
require this algorithm so it's low priority for now. Tracked by
NVIDIA/cub#285.