Increase floating point tolerance when validing results.
#295 reported that the `device_reduce` and `device_scan` tests were
failing due to the accumulated fp error exceeding the current threshold
of 0.01%. They observed a relative error of slightly more, 0.0100725%.

This appears to be happening on nvc++, though I've not been able to repro
the failure on NV HPC 21.7.

They suggest an alternate method of computing the summation that provides
more accuracy, but unfortunately this approach is not suitable for testing
cases where `scan_op` is not `cub::Sum`.

Instead, let's just bump up the error threshold to 0.015%.