Temporarily disable tests for non-RDC architectures.
This is due to how `CMAKE_CUDA_FLAGS` works -- all targets in a directory
must share the same flags. Now that some tests always enable RDC to test
CDP launches, we have to disable the tests for non-RDC arches (53, 62, 72).

We can fix this once we migrate to using `CMAKE_CUDA_ARCHITECTURES`.

See comments in this patch for details.