Better error message for no GPU or incompatible GPU
To give user some clue what's happening if the program
gets compiled on a node with no GPU or if it gets compiled
with different compute capability than the one it's running on.
In both scenarios no good error message were produced before.
The proposed changes will improve the user experience and make
it easier for users to troubleshoot problems.

This fix is for addressing the issue#1785 reported on Thrust
https://github.com/NVIDIA/thrust/issues/1785